plan_your_trip/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
appdirs==1.4.4
attrs==23.2.0
Automat==22.10.0
bcrypt==4.1.3
boto3==1.34.135
botocore==1.34.135
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==15.1.0
cryptography==42.0.8
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.15.4
frozenlist==1.4.1
greenlet==3.0.3
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
lxml==5.2.2
multidict==6.0.5
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parsel==1.9.1
pathlib==1.0.1
pip==24.0
plotly==5.22.0
Protego==0.3.1
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.5
pyOpenSSL==24.0.0
PySocks==1.7.1
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
queuelib==1.7.0
requests==2.32.3
requests-cache==0.4.13
requests-file==2.1.0
s3transfer==0.10.2
Scrapy==2.11.2
service-identity==21.1.0
setuptools==70.1.1
six==1.16.0
SQLAlchemy==2.0.31
tenacity==8.4.2
tldextract==5.1.2
tornado==6.4.1
Twisted==24.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
w3lib==2.2.1
wheel==0.43.0
yarl==1.9.4
zope.interface==6.4.post2
zstandard==0.22.0
